requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
peft>=0.7.0  # LoRA fine-tuning
bitsandbytes>=0.41.0  # Quantization (optional)

# Image Processing
Pillow>=10.0.0
pydicom>=2.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Azure
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
openai>=1.40.0  # For Azure OpenAI client

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
ipython>=8.0.0
jupyter>=1.0.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0
